Fix update after running tests #51
A few parts of the API have changed. In particular, weighted average is no longer available, and all 'blanks' are now set to None. Ex. 0 ratings or 0 ABV is now None, rather than a 0.

Additionally, a new exception JSONParseException has been added to make sure that when something strange comes down the pipes, we have something to catch it.